moves components broadly concerned with constructing the admin panel to an interface folder
