--- packages/cms/src/components/CtxMenu.css
@@ -1,1 +1,0 @@
-@import "../css/mixins.css";
