+++ packages/cms/src/components/interface/Header.css
@@ -1,4 +1,4 @@
-@import "../css/mixins.css";
+@import "../../css/mixins.css";
 
 .cms-header {
   display: flex;
