+++ packages/cms/src/components/interface/SelectorUsage.css
@@ -1,4 +1,4 @@
-@import "../css/mixins.css";
+@import "../../css/mixins.css";
 
 .cms-selector-usage.cms-card-container {
   padding-top: 0;
