+++ packages/cms/src/components/interface/SidebarTree.css
@@ -1,4 +1,4 @@
-@import "../css/mixins.css";
+@import "../../css/mixins.css";
 
 /* --------------------------------------- */
 /* TREE */
