+++ packages/cms/src/profile/ProfileBuilder.jsx
@@ -6,11 +6,11 @@ import ProfileEditor from "./ProfileEditor";
 import SectionEditor from "./SectionEditor";
 import PropTypes from "prop-types";
 import DimensionBuilder from "../profile/DimensionBuilder";
-import CtxMenu from "../components/CtxMenu";
 import Button from "../components/fields/Button";
-import SidebarTree from "../components/SidebarTree";
-import Header from "../components/Header";
-import Toolbox from "../components/toolbox/Toolbox";
+import CtxMenu from "../components/interface/CtxMenu";
+import SidebarTree from "../components/interface/SidebarTree";
+import Header from "../components/interface/Header";
+import Toolbox from "../components/interface/Toolbox";
 import nestedObjectAssign from "../utils/nestedObjectAssign";
 
 import varSwapRecursive from "../utils/varSwapRecursive";
