+++ packages/cms/src/story/StoryBuilder.jsx
@@ -2,9 +2,9 @@ import axios from "axios";
 import React, {Component} from "react";
 import {NonIdealState, Alert, Intent} from "@blueprintjs/core";
 import PropTypes from "prop-types";
-import CtxMenu from "../components/CtxMenu";
 import Button from "../components/fields/Button";
-import SidebarTree from "../components/SidebarTree";
+import CtxMenu from "../components/interface/CtxMenu";
+import SidebarTree from "../components/interface/SidebarTree";
 import StoryEditor from "./StoryEditor";
 import StorySectionEditor from "./StorySectionEditor";
