+++ packages/cms/src/story/StorySectionEditor.jsx
@@ -2,7 +2,7 @@ import axios from "axios";
 import React, {Component} from "react";
 
 import Button from "../components/fields/Button";
-import Accardion from "../components/Accardion";
+import Accardion from "../components/interface/Accardion";
 import TextCard from "../components/cards/TextCard";
 import Loading from "components/Loading";
 import VisualizationCard from "../components/cards/VisualizationCard";
